refactor: set cluster args to helm commands
