+++ src/utils/helm.ts
@@ -1,5 +1,10 @@
-import path from 'path';
+import {isEmpty} from 'lodash';
+import {dirname, join, sep} from 'path';
+import {v4 as uuid} from 'uuid';
 
+import {getHelmClusterArgs} from '@utils/cluster';
+
+import {CommandOptions, HelmEnv, HelmInstallArgs, HelmTemplateArgs} from '@shared/models/commands';
 import {HelmChart} from '@shared/models/helm';
 
 export function buildHelmCommand(
@@ -10,16 +15,16 @@ export function buildHelmCommand(
   rootFolderPath: string,
   clusterContext?: string
 ): string[] {
-  let chartFolderPath = path.join(rootFolderPath, path.dirname(helmChart.filePath));
+  let chartFolderPath = join(rootFolderPath, dirname(helmChart.filePath));
 
-  if (chartFolderPath.endsWith(path.sep)) {
+  if (chartFolderPath.endsWith(sep)) {
     chartFolderPath = chartFolderPath.slice(0, -1);
   }
 
   const args = [
     'helm',
     command,
-    ...valuesFilePaths.map(filePath => ['-f', `"${path.join(rootFolderPath, filePath)}"`]).flat(),
+    ...valuesFilePaths.map(filePath => ['-f', `"${join(rootFolderPath, filePath)}"`]).flat(),
     helmChart.name,
     `"${chartFolderPath}"`,
   ];
@@ -39,5 +44,302 @@ export function buildHelmCommand(
     args.push('--dry-run');
   }
 
+  const clusterArgs = getHelmClusterArgs();
+  args.push(...clusterArgs);
+
   return args;
+}
+
+export function createHelmInstallCommand({values, name, chart}: HelmInstallArgs, env?: HelmEnv): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['install', ...clusterArgs, '-f', `"${values}"`, chart, `"${name}"`, '--dry-run'],
+    env,
+  };
+}
+
+export function createHelmTemplateCommand({values, name, chart}: HelmTemplateArgs, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['template', '-f', `"${values}"`, chart, `"${name}"`],
+    env,
+  };
+}
+
+export function listHelmRepoCommand(env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['repo', 'list', '-o json'],
+    env,
+  };
+}
+
+export function addHelmRepoCommand({name, url}: {name: string; url: string}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['repo', 'add', name, url],
+    env,
+  };
+}
+
+export function updateHelmRepoCommand({repos = []}: {repos?: string[]}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['repo', 'update', ...repos],
+    env,
+  };
+}
+
+export function removeHelmRepoCommand({repos}: {repos: string[]}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['repo', 'remove', ...repos],
+    env,
+  };
+}
+
+export function indexHelmRepoCommand({dir}: {dir: string}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['repo', 'index', dir],
+    env,
+  };
+}
+
+export function searchHelmRepoCommand({q}: {q: string}, versions?: boolean, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['search', 'repo', q, versions ? '--versions' : '', '-o json'],
+    env,
+  };
+}
+
+export function searchHelmHubCommand({q}: {q: string}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['search', 'hub', q, '-o json'],
+    env,
+  };
+}
+
+export function helmChartInfoCommand({name}: {name: string}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['show', 'chart', name],
+    env,
+  };
+}
+
+export function helmChartReadmeCommand({name}: {name: string}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['show', 'readme', name],
+    env,
+  };
+}
+
+export function helmChartTemplateCommand({name}: {name: string}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['template', name],
+    env,
+  };
+}
+
+export function helmChartValuesCommand({name}: {name: string}, env?: HelmEnv): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['show', 'values', name],
+    env,
+  };
+}
+
+export function helmPullChartCommand(
+  {name, path, version}: {name: string; path: string; version?: string},
+  env?: HelmEnv
+): CommandOptions {
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['pull', name, '--untar', '--untardir', path, version ? `--version ${version}` : ''],
+    env,
+  };
+}
+
+export function installHelmRepoChartCommand(
+  {
+    chart,
+    namespace,
+    version,
+    shouldCreateNamespace,
+  }: {chart: string; namespace?: string; version?: string; shouldCreateNamespace?: boolean},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: [
+      'install',
+      chart,
+      '-g',
+      namespace ? `--namespace ${namespace}` : '',
+      version ? `--version ${version}` : '',
+      shouldCreateNamespace ? '--create-namespace' : '',
+      ...clusterArgs,
+    ],
+    env,
+  };
+}
+
+export function listHelmReleasesCommand(
+  {namespace, filter}: {namespace?: string; filter?: string},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: [
+      'list',
+      ...clusterArgs,
+      !isEmpty(namespace) ? `--namespace ${namespace}` : '--all-namespaces',
+      !isEmpty(filter) ? `--filter ${filter}` : '',
+      '-o json',
+    ],
+    env,
+  };
+}
+
+export function helmReleaseRevisionsCommand(
+  {release, namespace}: {release: string; namespace: string},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['history', release, ...clusterArgs, '-n', namespace, '-o json'],
+    env,
+  };
+}
+
+export function upgradeHelmReleaseCommand(
+  {
+    release,
+    chart,
+    namespace,
+    version,
+    dryRun,
+  }: {release: string; chart: string; namespace: string; version?: string; dryRun?: boolean},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: [
+      'upgrade',
+      release,
+      chart,
+      dryRun ? '--dry-run' : '',
+      ...clusterArgs,
+      '-n',
+      namespace,
+      version ? `--version ${version}` : '',
+    ],
+    env,
+  };
+}
+
+export function uninstallHelmReleaseCommand(
+  {release, namespace, dryRun}: {release: string; namespace: string; dryRun?: boolean},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['uninstall', release, dryRun ? '--dry-run' : '', ...clusterArgs, '-n', namespace],
+    env,
+  };
+}
+
+export function getHelmReleaseManifestCommand(
+  {release, namespace, revision = -1}: {release: string; namespace: string; revision?: number},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['get', 'manifest', release, revision > -1 ? `--revision ${revision}` : '', ...clusterArgs, '-n', namespace],
+    env,
+  };
+}
+
+export function getHelmReleaseValuesCommand(
+  {release, namespace}: {release: string; namespace: string},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['get', 'values', release, ...clusterArgs, '-n', namespace],
+    env,
+  };
+}
+
+export function getHelmReleaseNotesCommand(
+  {release, namespace}: {release: string; namespace: string},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['get', 'notes', release, ...clusterArgs, '-n', namespace],
+    env,
+  };
+}
+
+export function getHelmReleaseHooksCommand(
+  {release, namespace}: {release: string; namespace: string},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['get', 'hooks', release, ...clusterArgs, '-n', namespace],
+    env,
+  };
+}
+export function rollbackHelmReleaseCommand(
+  {release, namespace, revision}: {release: string; namespace: string; revision: number},
+  env?: HelmEnv
+): CommandOptions {
+  const clusterArgs = getHelmClusterArgs();
+  return {
+    commandId: uuid(),
+    cmd: 'helm',
+    args: ['rollback', release, revision.toString(), ...clusterArgs, '-n', namespace],
+    env,
+  };
 }
